Generify DataStore for Entities
